Fixed TestLogic1.java
I changed it back to what I changed it to:

A previous commit description: I know it says to not edit it, but when I tried to use this to check my other code, it didn't work since old35 and all the other functions were being called like a static function, even though they need an object. I just made it so I could test it by putting testLogic. before each function.